+++ packages/detect_proxy_setting/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.0.4
+* check proxy via url
+
 ## 0.0.3
 * add android