+++ packages/proxy_setting_android/CHANGELOG.md
@@ -1,2 +1,5 @@
+## 0.0.4
+* check proxy via url
+
 ## 0.0.3
 Add android