+++ packages/proxy_setting_platform_interface/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.0.3
+* fix description
+
 ## 0.0.2
 * add iOS